MAINT: Refactor and made changes based on review.
This commit moves the clever decorator by @pelson to the iris tests
package.
